Consolidate CardAction tests
